requirements.txt
Django==2.2.28
pytz==2022.1
sqlparse==0.4.2
python-decouple==3.6
gunicorn==20.1.0
dj-database-url==0.5.0
psycopg2-binary==2.8.6
boto3==1.24.4 
botocore==1.27.4 
django-s3-folder-storage==0.5 
django-storages==1.12.3 
jmespath==1.0.0 
python-dateutil==2.8.2 
s3transfer==0.6.0 
six==1.16.0 
urllib3==1.26.9
Collectfast==2.2.0 
typing-extensions==4.2.0
Django==4.0.5 
asgiref==3.5.2 
django-debug-toolbar==3.4.0
certifi==2022.5.18.1 
sentry-sdk==1.5.12